Add support to delay rendering in fallback funcs
Adds two driver functions:

set_pixel_raw: Draws a pixel without requiring the screen to render.
This funciton is optional and is not exposed through the high level
driver.  It is currently used by the fallback drawing routines.

render:  Render all changes to the lcd.

Calling setpixel from fallback functions like drawline, drawcircle,
etc.. can be slow on some displays.  For example, on the SH1106 the
setpixel generates a packet 4 bytes in size.  Drawing a filled rectangle
is viewable by the user.  With this patch the screen updates much
faster by only requiring 3 command packets followed by a stream of
pixels that have changed.

Note this does not break backwards functionality and doesn't force all
drivers to support this feature.